eliminar sesiones de un usuario antes de eliminarlo

+++ app/users/application/DeleteUser.py
@@ -3,6 +3,7 @@ from flask import abort
 from app.users.domain.UserId import UserId
 from app.users.domain.UserStructure import UserStructure
 from app.users.interface.FirestoreRepository import UserRepository
+from app.sessions.interface.FirestoreRepository import SessionRepository
 
 
 class DeleteUser(UserStructure):
@@ -11,5 +12,11 @@ class DeleteUser(UserStructure):
         if user_id.is_valid() is False:
             abort(400)
 
-        repo = UserRepository()
-        return repo.deleteUser(user_id.value)
+        session_repository = SessionRepository()
+        result = session_repository.delete_by_user_id(user_id.value)
+
+        if result:
+            repo = UserRepository()
+            return repo.deleteUser(user_id.value)
+        else:
+            abort(400)
